change synonym to word explorer
