+++ libs/scholars-scraper/src/lib/scraper/scholars-scraper.ts
@@ -1,7 +1,11 @@
-import { ParsedArticleHead } from '@foodmedicine/interfaces';
+import {
+  ParsedArticleHead,
+  ScholarsParserOpts,
+} from '@foodmedicine/interfaces';
 // eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
 import { Scraper } from '@foodmedicine/scraper';
 import * as parsers from '../parsers';
+import { getSynonyms } from '@foodmedicine/word-explorer';
 
 /**
  * Construct the google scholars url which will be scraped
@@ -11,7 +15,7 @@ function createScholarsUrl(
   impacted: string,
   solution: string,
   pageSize: number,
-  synonym=true
+  synonym = true
 ): string {
   return encodeURI(
     `https://www.ebi.ac.uk/europepmc/webservices/rest/search?query=${solution} ${impacted}&synonym=${synonym}&pageSize=${pageSize}`
@@ -26,7 +30,7 @@ function createScholarsUrl(
 export async function runScholarsScraper(
   impacted: string,
   recommendation: string,
-  pageSize = 25,
+  pageSize = 25
 ): Promise<ParsedArticleHead[]> {
   const queryUrl = createScholarsUrl(impacted, recommendation, pageSize);
   const remedyScraper = new Scraper<ParsedArticleHead>(parsers.ScholarsParser, {
@@ -35,7 +39,8 @@ export async function runScholarsScraper(
       recommendation: recommendation,
       impacted: impacted,
     },
-  });
+    impactedSynonyms: await getSynonyms(impacted),
+  } as ScholarsParserOpts);
 
   const articleHeads = await remedyScraper.run();
   return articleHeads;
